Fix a bug within EGraphMatcher::all_captures
The symptom here was that we would fail an assert indicating that we
tried to call pop_back on an empty vector when attempting to remove a
capture.

The issue was that there was an old guard statement left over that was
deleting all the captures for the subclause indiscriminately. This was
correct when there was only 1 capture per clause, but now that there are
multiple it is wrong.