Implement a captures system for `EMatcher` (#729)
* Add support for captures to ematching subclauses

Currently, when we write a rewrite update function we need to iterate
over all the possible matches within the update function. This is
tedious and somewhat difficult to do correctly, especially as the
expression tree that is being matched gets more and more complicated.

To address this, I'm introducing the idea of a capture. This means that
the corresponding ENode can be easily accessed by the submatcher ID
without needing to manually iterate over the appropriate matches within
the update function for each and every rewrite.

The code to match and prepare the captures will be added in a future
commit.

* Modify EGraphMatcher to call update for each set of captures

This is done by recursively iterating through all possible combinations
of captures while ignoring any subtrees that are not captured as they
are not expected to matter.

* Rewrite the zext-trunc rewrite to use captures

* Add a test case verifying that multiple captures work

Specifically, this verifies that when there's multiple possible values
for the multiple captures everything still works correctly.

* Formatting