Allow captures within e-matching with multiple values (#730)
This PR is an extension to #729 that extends the captures system to allow a clause to capture multiple values simultaneously. Immediately after implementing #729, I found a use case that still didn't work (rewriting `(shl (lshr ?x ?y) ?z)` where `?y` and `?z` are constants) so I've now gone and extended capturing to support that use case as well.

* Add a typedef for the type of the captures map

This means that it can be easily refactored when necessary.

* Add support for subclauses with multiple captures

* Fix a bug within EGraphMatcher::all_captures

The symptom here was that we would fail an assert indicating that we
tried to call pop_back on an empty vector when attempting to remove a
capture.

The issue was that there was an old guard statement left over that was
deleting all the captures for the subclause indiscriminately. This was
correct when there was only 1 capture per clause, but now that there are
multiple it is wrong.