Rethink `MemHeap` and `MemHeapMgr` to be more maintainable (#740)
Strictly this is more-or-less a rewrite of `MemHeap` and `MemHeapMgr` to change two things:
- The internal concrete allocator now has a virtual interface instead of using a variant
- `allocate`, `deallocate`, and `resolve` now take an `InterpreterContext` instance

Since these changes break the interface I've created them in separate `Heap`/`MultiHeap`  classes which are currently unused. In a future PR I'll switch the rest of the codebase over to the new classes. Note that the implementation is mostly a copy-paste from `MemHeap` and `MemHeapMgr` with some small modifications to work with the two changes above.