Fix broken install for rules with include_prefix
It turns out that the packaging rules I wrote for headers didn't account
for rules which use include_prefix. This fixes the issue by adding a
prefix if one is specified.

As a side effect, this means that the immer headers are actually
installed in the right location within the install directory.