Implement a caching solver on top of LMDB (#751)
This is the simplest possible implementation of a caching solver built on top of LMDB. All it does is convert each expression in the solver to a string, sort them, and then use them as the cache key. LMDB isn't optimal for this since it has a compile-time upper bound on the length of a key but I've bumped that up somewhat so that it works okay enough for now.

In the future it would be better to use a kv-store that supports arbitrary-length keys but I haven't found a local one as-of yet. RocksDB would probably be a good option for this.